refactor: Move traj to proto track converter to separate target
