+++ Examples/Algorithms/CMakeLists.txt
@@ -15,3 +15,4 @@ add_subdirectory(TrackFittingChi2)
 add_subdirectory(TruthTracking)
 add_subdirectory(Vertexing)
 add_subdirectory_if(Alignment ACTS_BUILD_ALIGNMENT)
+add_subdirectory(Utilities)
