+++ Examples/Algorithms/TrackFinding/CMakeLists.txt
@@ -7,8 +7,7 @@ add_library(
   src/TrackFindingAlgorithm.cpp
   src/TrackFindingAlgorithmFunction.cpp
   src/HoughTransformSeeder.cpp
-  src/TrackParamsEstimationAlgorithm.cpp
-  src/TrajectoriesToPrototracks.cpp)
+  src/TrackParamsEstimationAlgorithm.cpp)
 target_include_directories(
   ActsExamplesTrackFinding
   PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
