+++ Examples/Algorithms/Utilities/CMakeLists.txt
@@ -1,0 +1,15 @@
+add_library(
+  ActsExamplesUtilities SHARED
+  src/TrajectoriesToPrototracks.cpp)
+target_include_directories(
+  ActsExamplesUtilities
+  PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
+target_link_libraries(
+  ActsExamplesUtilities
+  PUBLIC
+    ActsCore
+    ActsExamplesFramework)
+
+install(
+  TARGETS ActsExamplesUtilities
+  LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR})
